backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
idna==3.11
marshmallow==3.26.1
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-decouple==3.8
requests==2.33.1
sqlparse==0.5.5
transbank-sdk==6.1.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0
